Fix the playback pitch ceiling, ping-pong tuning and a short loop crash (#2665)
The playback rate topped out at 128 times the sample's natural pitch, so
a long single cycle stopped rising well inside the keyboard. A 2048
sample cycle stalled around 3 kHz. With zero order hold, which runs
without oversampling, notes near the limit played at the wrong pitch
entirely. The rate now has room to go past that.

A ping-pong loop turned around wherever the playhead happened to land
past the loop end instead of mirroring back from the end itself. The
loop played longer than it should, so short loops sounded flat, and by
uneven steps as the pitch rose. Ping-pong loops now reflect off both
ends at the exact loop bounds, even when one step crosses the loop more
than once, and still count a loop each time they return to where they
set out.

Looping a sample shorter than about 24 samples, like a small chiptune
wavetable cycle, read far outside the sample and could crash the
process. Short loops now wrap back into the loop and play at pitch.

The generator also gets its first tests.

Closes #2371

+++ src/scxt-core/dsp/generator.h
@@ -84,7 +84,7 @@ struct GeneratorState
     int32_t loopLowerBound{0};     // inclusive
     int32_t loopUpperBound{1};     // inclusive
     float loopInvertedBounds{1.f}; // 1 / (UB-LB)
-    int32_t ratio{1 << 24};        // 1 << 24 is playback-at-tempo
+    int64_t ratio{1 << 24};        // 1 << 24 is playback-at-tempo
     int16_t blockSize{scxt::blockSize};
     bool isFinished{true};
     int32_t sampleStart{0};
